Initial commit with some clean up unused code
Initial commit with some clean up unused code
