+++ backend/main.py
@@ -1,7 +1,7 @@
 import logging
 from fastapi import FastAPI, Request, APIRouter
 from fastapi.middleware.cors import CORSMiddleware
-from loader_scheduler import LoaderScheduler
+from backend.loader_service import LoaderScheduler
 import threading
 
 # Configure logging
@@ -27,70 +27,3 @@ router = APIRouter()
 async def read_root(request: Request):
     return {"message": "Server is running"}
 
-@app.post("/scheduler_overview")
-async def scheduler_overview():
-    try:
-        overview = str(scheduler.scheduler)
-        overview_lines = overview.split("\n")
-        overview_dict = {
-            "max_exec": overview_lines[0].split(",")[0].split("=")[1].strip(),
-            "tzinfo": overview_lines[0].split(",")[1].split("=")[1].strip(),
-            "priority_function": overview_lines[0].split(",")[2].split("=")[1].strip(),
-            "jobs": []
-        }
-        for line in overview_lines[3:]:
-            if line.strip() and not line.startswith("--------"):
-                parts = line.split()
-                if len(parts) >= 8:
-                    job = {
-                        "type": parts[0],
-                        "function": parts[1],
-                        "due_at": f"{parts[2]} {parts[3]}",
-                        "tzinfo": parts[4],
-                        "due_in": parts[5]
-                    }
-                    # Replace function names
-                    if job["function"] == "#et_data_etl(..)":
-                        job["function"] = "run_yfinance_market_data_etl"
-                    elif job["function"] == "#ic_data_etl(..)":
-                        job["function"] = "run_pyfredapi_macroeconomic_data_etl"
-                    elif job["function"] == "#_processing(..)":
-                        job["function"] = "run_financial_news_processing"
-                    
-                    # Add "d" to single digit due_in values
-                    if job["due_in"].isdigit():
-                        job["due_in"] += "d"
-                    
-                    overview_dict["jobs"].append(job)
-                else:
-                    job = {
-                        "type": parts[0] if len(parts) > 0 else "",
-                        "function": parts[1] if len(parts) > 1 else "",
-                        "due_at": f"{parts[2]} {parts[3]}" if len(parts) > 3 else "",
-                        "tzinfo": parts[4] if len(parts) > 4 else "",
-                        "due_in": parts[5] if len(parts) > 5 else ""
-                    }
-                    # Replace function names
-                    if job["function"] == "#et_data_etl(..)":
-                        job["function"] = "run_yfinance_market_data_etl"
-                    elif job["function"] == "#ic_data_etl(..)":
-                        job["function"] = "run_pyfredapi_macroeconomic_data_etl"
-                    elif job["function"] == "#_processing(..)":
-                        job["function"] = "run_financial_news_processing"
-                    
-                    # Add "d" to single digit due_in values
-                    if job["due_in"].isdigit():
-                        job["due_in"] += "d"
-                    
-                    overview_dict["jobs"].append(job)
-        return {"overview": overview_dict}
-    except Exception as e:
-        logger.error(f"Error generating scheduler overview: {e}")
-        return {"error": "Failed to generate scheduler overview"}
-
-def start_scheduler():
-    scheduler.start()
-
-scheduler = LoaderScheduler()
-scheduler_thread = threading.Thread(target=start_scheduler)
-scheduler_thread.start()
