+++ backend/pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "your-virtual-env"
 version = "0.1.0"
-description = "This repository contains the backend implementation of data loaders for the Capital Markets demo."
+description = "This repository hosts the backend for the Capital Markets Agents service."
 authors = ["Julian Boronat <julian.boronat@mongodb.com>"]
 readme = "README.md"
 
@@ -13,7 +13,6 @@ fastapi = "^0.115.4"
 uvicorn = "^0.32.0"
 boto3 = "^1.35.70"
 botocore = "^1.35.70"
-scheduler = "^0.8.8"
 tqdm = "^4.67.1"
 transformers = "^4.49.0"
 torch = "^2.6.0"
