+++ docker-compose.yml
@@ -1,11 +1,11 @@
 services:
-  capitalmarkets-loaders:
+  capitalmarkets-agents:
     build:
       context: .
       no_cache: true
       dockerfile: Dockerfile.backend
-    image: capitalmarkets-loaders-image:latest
+    image: capitalmarkets-agents-image:latest
     ports:
       - "8000:8000"
     restart: always
-    container_name: capitalmarkets-loaders-container
+    container_name: capitalmarkets-agents-container
