--- notebooks/csv/enriched/Macroeconomic_Indicator_GDP.csv
@@ -1,13 +1,0 @@
-title,series_id,frequency,frequency_short,units,units_short,date,value
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2022-01-01,25215.491
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2022-04-01,25805.791
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2022-07-01,26272.011
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2022-10-01,26734.277
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2023-01-01,27164.359
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2023-04-01,27453.815
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2023-07-01,27967.697
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2023-10-01,28296.967
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2024-01-01,28624.069
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2024-04-01,29016.714
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2024-07-01,29374.914
-Gross Domestic Product,GDP,Quarterly,Q,Billions of Dollars,Bil. of $,2024-10-01,29719.647
