--- notebooks/csv/enriched/Macroeconomic_Indicator_REAINTRATREARAT10Y.csv
@@ -1,39 +1,0 @@
-title,series_id,frequency,frequency_short,units,units_short,date,value
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-01-01,0.3194951
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-02-01,0.44159289
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-03-01,0.34708992
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-04-01,0.94940027
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-05-01,1.20389776
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-06-01,1.02509351
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-07-01,1.06217493
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-08-01,0.8556253
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-09-01,1.20424868
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-10-01,1.79733203
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-11-01,1.93071305
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2022-12-01,1.57612515
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-01-01,1.78309897
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-02-01,1.42020217
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-03-01,2.05963636
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-04-01,1.44391386
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-05-01,1.53690406
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-06-01,1.06063107
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-07-01,1.4260473
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-08-01,1.60209539
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-09-01,1.69970754
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-10-01,2.08246076
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-11-01,2.09432864
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2023-12-01,1.68089907
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-01-01,1.68087145
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-02-01,1.6168231
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-03-01,1.92596928
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-04-01,1.93819396
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-05-01,2.10339878
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-06-01,1.99956671
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-07-01,2.04839593
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-08-01,1.66418984
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-09-01,1.58158069
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-10-01,1.48014834
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-11-01,1.95841688
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2024-12-01,1.82436879
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2025-01-01,2.05553249
-10-Year Real Interest Rate,REAINTRATREARAT10Y,Monthly,M,Percent,%,2025-02-01,2.03128216
