--- notebooks/csv/enriched/Macroeconomic_Indicator_UNRATE.csv
@@ -1,38 +1,0 @@
-title,series_id,frequency,frequency_short,units,units_short,date,value
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-01-01,4
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-02-01,3.8
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-03-01,3.7
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-04-01,3.7
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-05-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-06-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-07-01,3.5
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-08-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-09-01,3.5
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-10-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-11-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2022-12-01,3.5
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-01-01,3.5
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-02-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-03-01,3.5
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-04-01,3.4
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-05-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-06-01,3.6
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-07-01,3.5
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-08-01,3.7
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-09-01,3.8
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-10-01,3.9
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-11-01,3.7
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2023-12-01,3.8
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-01-01,3.7
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-02-01,3.9
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-03-01,3.9
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-04-01,3.9
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-05-01,4
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-06-01,4.1
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-07-01,4.2
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-08-01,4.2
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-09-01,4.1
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-10-01,4.1
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-11-01,4.2
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2024-12-01,4.1
-Unemployment Rate,UNRATE,Monthly,M,Percent,%,2025-01-01,4
