--- notebooks/csv/raw/gdp_index_historical.csv
@@ -1,13 +1,0 @@
-realtime_start,realtime_end,date,value
-2025-03-04,2025-03-04,2022-01-01,25215.491
-2025-03-04,2025-03-04,2022-04-01,25805.791
-2025-03-04,2025-03-04,2022-07-01,26272.011
-2025-03-04,2025-03-04,2022-10-01,26734.277
-2025-03-04,2025-03-04,2023-01-01,27164.359
-2025-03-04,2025-03-04,2023-04-01,27453.815
-2025-03-04,2025-03-04,2023-07-01,27967.697
-2025-03-04,2025-03-04,2023-10-01,28296.967
-2025-03-04,2025-03-04,2024-01-01,28624.069
-2025-03-04,2025-03-04,2024-04-01,29016.714
-2025-03-04,2025-03-04,2024-07-01,29374.914
-2025-03-04,2025-03-04,2024-10-01,29719.647
