--- notebooks/csv/raw/real_interest_rate_historical.csv
@@ -1,39 +1,0 @@
-realtime_start,realtime_end,date,value
-2025-03-04,2025-03-04,2022-01-01,0.3194951
-2025-03-04,2025-03-04,2022-02-01,0.44159289
-2025-03-04,2025-03-04,2022-03-01,0.34708992
-2025-03-04,2025-03-04,2022-04-01,0.94940027
-2025-03-04,2025-03-04,2022-05-01,1.20389776
-2025-03-04,2025-03-04,2022-06-01,1.02509351
-2025-03-04,2025-03-04,2022-07-01,1.06217493
-2025-03-04,2025-03-04,2022-08-01,0.8556253
-2025-03-04,2025-03-04,2022-09-01,1.20424868
-2025-03-04,2025-03-04,2022-10-01,1.79733203
-2025-03-04,2025-03-04,2022-11-01,1.93071305
-2025-03-04,2025-03-04,2022-12-01,1.57612515
-2025-03-04,2025-03-04,2023-01-01,1.78309897
-2025-03-04,2025-03-04,2023-02-01,1.42020217
-2025-03-04,2025-03-04,2023-03-01,2.05963636
-2025-03-04,2025-03-04,2023-04-01,1.44391386
-2025-03-04,2025-03-04,2023-05-01,1.53690406
-2025-03-04,2025-03-04,2023-06-01,1.06063107
-2025-03-04,2025-03-04,2023-07-01,1.4260473
-2025-03-04,2025-03-04,2023-08-01,1.60209539
-2025-03-04,2025-03-04,2023-09-01,1.69970754
-2025-03-04,2025-03-04,2023-10-01,2.08246076
-2025-03-04,2025-03-04,2023-11-01,2.09432864
-2025-03-04,2025-03-04,2023-12-01,1.68089907
-2025-03-04,2025-03-04,2024-01-01,1.68087145
-2025-03-04,2025-03-04,2024-02-01,1.6168231
-2025-03-04,2025-03-04,2024-03-01,1.92596928
-2025-03-04,2025-03-04,2024-04-01,1.93819396
-2025-03-04,2025-03-04,2024-05-01,2.10339878
-2025-03-04,2025-03-04,2024-06-01,1.99956671
-2025-03-04,2025-03-04,2024-07-01,2.04839593
-2025-03-04,2025-03-04,2024-08-01,1.66418984
-2025-03-04,2025-03-04,2024-09-01,1.58158069
-2025-03-04,2025-03-04,2024-10-01,1.48014834
-2025-03-04,2025-03-04,2024-11-01,1.95841688
-2025-03-04,2025-03-04,2024-12-01,1.82436879
-2025-03-04,2025-03-04,2025-01-01,2.05553249
-2025-03-04,2025-03-04,2025-02-01,2.03128216
